Move blobCacheSource into a separate source file
... to make the files a bit shorter.

Only moves the code without modification, should not
change behavior.
